requirements.txt
requests>=2.28.1
flask>=2.3.0
flask-cors>=4.0.0
pyjwt>=2.4.0
werkzeug>=2.3.0
python-dateutil>=2.8.2
sqlalchemy>=2.0.0
flask-sqlalchemy>=3.0.0
flask-mail>=0.9.1
itsdangerous>=2.1.0
